+++ code/test/client.c
@@ -62,9 +62,9 @@ void doTest(){
 int main(){
     int i = 0;
 
-    for(i = 0; i < 10; i++){
+    for(i = 0; i < 3; i++){
         doTest();
     }
 
-    
+    Halt();
 }
